refactor monitoring RPCs
fix operations monitor rpc - add protocol and better error handling
